moved photo details components to dir
