+++ src/components/ImageDetails/ImageInfo.vue
