+++ src/components/ImageDetails/LegalDisclaimer.vue
