+++ src/components/ImageDetails/ReuseSurvey.vue
@@ -37,6 +37,3 @@ export default {
   },
 };
 </script>
-
-<style lang="scss" scoped>
-</style>
